Worked on a lot of refactoring
