--- FrontEnd/src/components/ButtonComponent/Button.css
@@ -1,50 +1,0 @@
-@import url('https://fonts.googleapis.com/css2?family=Roboto:wght@500&display=swap');
-
-a.btn{
-    --primary: rgb(136,0,21);
-    text-decoration: none;
-    text-align: center;
-    padding: 8px 20px;
-    border-radius: 4px;
-    font-family: 'Roboto';
-    outline: none;
-    border: none;
-    cursor: pointer;
-    color: #ffffff;
-}
-
-a.btn:hover{
-    padding: 8px 20px;
-    transition: all 0.3s ease-out;
-    background: #ffffff;
-    color: rgb(136,0,21);
-    transition: 250ms;
-}
-
-a.btn--primary{
-    background-color: var(--primary);
-}
-
-a.btn-outline{
-    background-color: transparent;
-    color: #ffffff;
-    padding: 8px 20px;
-    border-radius: 4px;
-    border: 1px solid var(--primary);
-    transition: all 0.3s ease-out;
-}
-
-a.btn-medium{
-    padding: 8px 20px;
-    border-radius: 4px;
-    font-size: 18px;
-    color: #ffffff;
-}
-
-a.btn-large{
-    padding: 12px 28px;
-    border-radius: 4px;
-    font-size: 20px;
-    color: #ffffff;
-
-}
