+++ FrontEnd/src/components/DisplayLogosComponent/DisplayLogos.tsx
@@ -2,15 +2,19 @@
 
 import React, { useState, useEffect } from 'react';
 import { CSSTransition } from 'react-transition-group';
+import Fade from '../util/FadeComponent/Fade';
 import './DisplayLogos.css';
 
-const DisplayLogos = ({
-    images, text, staticImage, backgroundColor}:{
+
+
+const DisplayLogos = ({images, text, staticImage, backgroundColor, numImages}:{
         images:Array<string>;
         text:string;
         staticImage:boolean;
         backgroundColor:string;
+        numImages: number;
     }) => {
+    const nImages = (typeof(numImages) == 'number' ? numImages : 1);
     const staticImg = (typeof(staticImage) == 'boolean' ? staticImage : false);
     const bgColor = (typeof(backgroundColor) == 'string' ? backgroundColor: "#888888");
     //states
@@ -22,34 +26,6 @@ const DisplayLogos = ({
     const [intermission, setIntermission] = useState(false);
     //functions
     const arrayColumn = (arr:any, n:any) => arr.map((x:any) => x[n]);
-    const cssTransitionFunction = (n:number) => {
-        if(data.length !== 0){
-            return(
-                <CSSTransition 
-                    in={animationState}
-                    timeout={500}
-                    unmountOnExit
-                    classNames = 'img'
-                    >
-                        <img src = {data[(current_image + n) % data.length]} className = "img-logo" alt = "opps"/>
-                </CSSTransition> 
-            );
-        }
-    };
-    const dynamicImageFunction = () => {
-        return(
-            <div>
-                {cssTransitionFunction(0)}
-                {cssTransitionFunction(1)}
-                {cssTransitionFunction(2)}
-                {cssTransitionFunction(3)}
-                {cssTransitionFunction(4)}
-                {cssTransitionFunction(5)}
-                {cssTransitionFunction(6)}
-                {cssTransitionFunction(7)}
-            </div>
-        )
-    }
     //variables
     var updateData = false;
 
@@ -70,42 +46,8 @@ const DisplayLogos = ({
         }
         //fetchingAPI();
         setData(images);
-        setFadeIn(true);
 
     }, [updateData]);
-
-    useEffect(() => {
-        if (data.length !== 0){
-            imageSwitch((current_image + 8) % data.length);
-            setAnimationState(!animationState);
-            setFadeOut(!fadeOut);
-
-        }
-    }, [fadeIn, setFadeIn]);
-
-    useEffect(() => {
-        if (data.length !== 0){
-            var imageTimeout =  setTimeout(() => {
-                setAnimationState(!animationState);
-                setIntermission(!intermission);
-            }, 4500);
-            return () => {
-                clearTimeout(imageTimeout);
-            };
-        }
-    }, [fadeOut, setFadeOut]);
-
-    useEffect(() => {
-        if (data.length !== 0){
-            var imageTimeout = setTimeout(() => {
-                setFadeIn(!fadeIn);
-            }, 500);
-            return () => {
-                clearTimeout(imageTimeout);
-            };
-        }
-    }, [intermission, setIntermission]);
-
     //rendering
     if (data.length === 0){
         updateData = !updateData;
@@ -115,7 +57,7 @@ const DisplayLogos = ({
             </div>
         );
     }
-
+    console.log(images);
     return(   
         <div className = "speaker-container" style= {{background: bgColor}}>
             <h1 className = "speaker-header"><strong>{text}</strong></h1>
@@ -129,7 +71,7 @@ const DisplayLogos = ({
                         )
                     })
                     :
-                    dynamicImageFunction()
+                    <Fade data={images} numImages={nImages}/>
                 }
 
             </div>
